:boom: 后端重构完成

+++ spark-yun-backend/src/main/java/com/isxcode/star/backend/module/engine/node/service/EngineNodeBizService.java
@@ -3,6 +3,7 @@ package com.isxcode.star.backend.module.engine.node.service;
 import static com.isxcode.star.backend.utils.SshUtils.executeCommand;
 import static com.isxcode.star.backend.utils.SshUtils.scpFile;
 
+import com.isxcode.star.api.constants.EngineNodeStatus;
 import com.isxcode.star.api.constants.PathConstants;
 import com.isxcode.star.api.pojos.engine.node.req.EnoAddNodeReq;
 import com.isxcode.star.api.pojos.engine.node.req.EnoQueryNodeReq;
@@ -18,11 +19,15 @@ import com.isxcode.star.backend.module.engine.node.mapper.EngineNodeMapper;
 import com.isxcode.star.backend.module.engine.node.repository.EngineNodeRepository;
 import com.isxcode.star.common.exception.SparkYunException;
 import java.io.File;
+import java.io.IOException;
 import java.util.Optional;
 import javax.transaction.Transactional;
+
+import com.jcraft.jsch.JSchException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.logging.log4j.util.Strings;
+import org.mapstruct.Mapping;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -61,6 +66,9 @@ public class EngineNodeBizService {
       node.setAgentPort(sparkYunProperties.getDefaultAgentPort());
     }
 
+    // 初始化节点状态，未检测
+    node.setStatus(EngineNodeStatus.UN_CHECK);
+
     engineNodeRepository.save(node);
   }
 
@@ -89,6 +97,16 @@ public class EngineNodeBizService {
   }
 
   public void delNode(String nodeId) {
+
+    Optional<EngineNodeEntity> engineNodeEntityOptional = engineNodeRepository.findById(nodeId);
+    if (!engineNodeEntityOptional.isPresent()) {
+      throw new SparkYunException("节点已删除");
+    }
+
+    // 判断节点状态是否为已安装
+    if (EngineNodeStatus.ACTIVE.equals(engineNodeEntityOptional.get().getStatus())) {
+      throw new SparkYunException("请卸载节点后删除");
+    }
 
     engineNodeRepository.deleteById(nodeId);
   }
@@ -128,10 +146,22 @@ public class EngineNodeBizService {
 
     // 运行安装脚本
     String installCommand =
-      "bash " + engineNode.getAgentHomePath() + File.separator + PathConstants.AGENT_INSTALL_BASH_NAME + " --home-path=" + engineNode.getAgentHomePath() + " --agent-port=" + engineNode.getAgentPort();
-    String executeLog = executeCommand(engineNode, installCommand, false);
+      "bash " + engineNode.getAgentHomePath() + File.separator + PathConstants.AGENT_INSTALL_BASH_NAME
+        + " --home-path=" + engineNode.getAgentHomePath()
+        + " --agent-port=" + engineNode.getAgentPort()
+        + " --hadoop-home=" + engineNode.getHadoopHomePath();
 
-    engineNode.setStatus("INSTALLED");
+    String executeLog = null;
+    try {
+      executeLog = executeCommand(engineNode, installCommand, false);
+    } catch (JSchException | InterruptedException | IOException e) {
+      log.error(e.getMessage());
+      engineNode.setStatus(EngineNodeStatus.INSTALL_ERROR);
+      engineNodeRepository.save(engineNode);
+      throw new SparkYunException("安装失败", e.getMessage());
+    }
+
+    engineNode.setStatus(EngineNodeStatus.ACTIVE);
     engineNodeRepository.save(engineNode);
 
     return new EnoInstallAgentRes(executeLog);
@@ -143,9 +173,17 @@ public class EngineNodeBizService {
 
     // 运行卸载脚本
     String checkCommand = "echo \"可安装\"";
-    String executeLog = executeCommand(engineNode, checkCommand, false);
+    String executeLog = null;
+    try {
+      executeLog = executeCommand(engineNode, checkCommand, false);
+    } catch (JSchException | InterruptedException | IOException e) {
+      log.error(e.getMessage());
+      engineNode.setStatus(EngineNodeStatus.CAN_NOT_INSTALL);
+      engineNodeRepository.save(engineNode);
+      throw new SparkYunException("不可安装", e.getMessage());
+    }
 
-    engineNode.setStatus("CAN_INSTALLED");
+    engineNode.setStatus(EngineNodeStatus.CAN_INSTALL);
     engineNodeRepository.save(engineNode);
 
     return new EnoCheckAgentRes(executeLog);
@@ -158,9 +196,17 @@ public class EngineNodeBizService {
     // 运行卸载脚本
     String installCommand =
       "bash " + engineNode.getAgentHomePath() + File.separator + "spark-yun-agent" + File.separator + "bin" + File.separator + PathConstants.AGENT_REMOVE_BASH_NAME + " --home-path=" + engineNode.getAgentHomePath();
-    String executeLog = executeCommand(engineNode, installCommand, false);
+    String executeLog = null;
+    try {
+      executeLog = executeCommand(engineNode, installCommand, false);
+    } catch (JSchException | InterruptedException | IOException e) {
+      log.error(e.getMessage());
+      engineNode.setStatus(EngineNodeStatus.ACTIVE);
+      engineNodeRepository.save(engineNode);
+      throw new SparkYunException("卸载失败", e.getMessage());
+    }
 
-    engineNode.setStatus("UNINSTALLED");
+    engineNode.setStatus(EngineNodeStatus.UNINSTALLED);
     engineNodeRepository.save(engineNode);
 
     return new EnoRemoveAgentRes(executeLog);
